Replace `logging` call with `print()`.
This is the only place in the entire package where we use `logging`
instead of `print()` for logging.  It is also the only place in this
package that prints to stdout instead of stdin which means these
messages don't always show up where they're supposed to…

I'd opt following for the path of least resistance and just go full
`print()`.